requirements.txt
# Core dependencies
python-dotenv==1.0.0
pypdf==3.17.0

# RAG Framework - use compatible versions
langchain==0.0.340
chromadb==0.4.18
sentence-transformers==2.2.2

# LLM - use latest compatible version
openai==0.28.0
# Removed anthropic for now - we'll add it later if needed

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1

# Utilities
numpy==1.24.3
pandas==2.1.3
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
